--- a/tfjs-backend-webgpu/src/reduce_webgpu.ts
+++ b/tfjs-backend-webgpu/src/reduce_webgpu.ts
@@ -33,12 +33,17 @@ export class ReduceProgram implements WebGPUProgram {
 
   constructor(
       reduceInfo: backend_util.ReduceInfo,
-      reduceType: 'all'|'any'|'max'|'mean'|'min'|'prod'|'sum') {
+      reduceType: 'all'|'any'|'max'|'mean'|'min'|'prod'|'sum',
+      maxComputeWorkgroupSizeX: number) {
     this.inputShape = [reduceInfo.batchSize, reduceInfo.inSize];
     const [outputShape, ] =
         backend_util.computeOutAndReduceShapes(this.inputShape, [1]);
     this.outputShape = outputShape.length === 0 ? [1] : outputShape;
-
+    if (reduceInfo.inSize >= 32768 && maxComputeWorkgroupSizeX >= 512) {
+      this.workgroupSize = [512, 1, 1];
+    } else if (reduceInfo.inSize >= 4096) {
+      this.workgroupSize = [256, 1, 1];
+    }
     this.dispatchLayout = flatDispatchLayout(this.outputShape);
     // A work group only outputs a data, so we transfer [1, 1, 1] to compute
     // dispatch size.