PERF

Increase the workgroupSize when the reduce length is very large. See great improvement of Mean in stable diffusion. Mean in stable diffusion: 16ms -> 8ms
Mean in decoder: 165ms -> 35.92ms

To see the logs from the Cloud Build CI, please join either our [discussion](https://groups.google.com/a/tensorflow.org/forum/#!forum/tfjs) or [announcement](https://groups.google.com/a/tensorflow.org/forum/#!forum/tfjs-announce) mailing list.

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/tensorflow/tfjs/7559)
<!-- Reviewable:end -->